Added subscribing and unsubscribing

+++ InstagramHelp/Models/Account/UserData.cs
@@ -3,7 +3,7 @@ namespace InstagramHelp.Models.Account
     public sealed record UserData(
         string UserName,
         string FullName,
-        int UserMediaAmount,
-        int FollowersAmount,
-        int FollowingAmount);
+        long UserMediaAmount,
+        long FollowersAmount,
+        long FollowingAmount);
 }
